22 ФЗ 1 правки

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/OrderDto.java
@@ -1,18 +1,22 @@
 package ru.yandex.practicum.dto;
 
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 import ru.yandex.practicum.constant.OrderState;
 
 import java.util.Map;
 
-@Getter @Setter
+@Getter
+@Setter
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class OrderDto {
     String orderId;
     String shoppingCartId;
-    Map<String, Integer> products;
+    Map<String, Long> products;
     String paymentId;
     String deliveryId;
     OrderState state;
