+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/ShoppingCartDto.java
@@ -10,5 +10,5 @@ import java.util.Map;
 public class ShoppingCartDto {
     private String username;
     private String shoppingCartId;
-    private Map<String, Integer> products;
+    private Map<String, Long> products;
 }
