+++ commerce/order/src/main/java/ru/yandex/practicum/model/Order.java
@@ -38,7 +38,7 @@ public class Order {
     @MapKeyColumn(name = "product_id")
     @Column(name = "quantity")
     @Builder.Default
-    Map<UUID, Integer> products = new HashMap<>();
+    Map<UUID, Long> products = new HashMap<>();
 
     UUID paymentId;
     UUID deliveryId;
